+++ test/unittests/test_git_cache_database.py
@@ -13,6 +13,7 @@
 # Module Import
 # -----------------------------------------------------------------------------
 import importlib
+import json
 import os
 import time
 from unittest import TestCase
@@ -61,51 +62,58 @@ class GitCacheDatabaseTest(TestCase):
         database = git_cache.database.Database()
         self.assertFalse(database.database)
 
-        database.add("http://dummy/git", "dummy-dir")
-        self.assertIn("dummy-dir", database.get_all())
-        self.assertEqual("http://dummy/git", database.get("dummy-dir")["url"])
-        self.assertNotEqual(0, database.get("dummy-dir")["last-update-time"])
-        self.assertEqual(0, database.get("dummy-dir")["mirror-updates"])
-        self.assertEqual(0, database.get("dummy-dir")["clones"])
-        self.assertEqual(0, database.get("dummy-dir")["updates"])
+        repo_abs_path = os.path.normpath(os.path.join("/tmp", "dummy-dir"))
+        database.add("http://dummy/git", repo_abs_path)
+        self.assertIn(repo_abs_path, database.get_all())
+        self.assertEqual("http://dummy/git", database.get(repo_abs_path)["url"])
+        self.assertNotEqual(0, database.get(repo_abs_path)["last-update-time"])
+        self.assertEqual(0, database.get(repo_abs_path)["mirror-updates"])
+        self.assertEqual(0, database.get(repo_abs_path)["clones"])
+        self.assertEqual(0, database.get(repo_abs_path)["updates"])
 
-        database.increment_counter("dummy-dir", "mirror-updates")
-        self.assertEqual(1, database.get("dummy-dir")["mirror-updates"])
-        self.assertEqual(0, database.get("dummy-dir")["clones"])
-        self.assertEqual(0, database.get("dummy-dir")["updates"])
+        # Ensure written database uses relative paths
+        with open("/tmp/db", "r", encoding="utf-8") as handle:
+            raw_database = json.load(handle)
+        self.assertNotIn(repo_abs_path, raw_database)
+        self.assertIn("dummy-dir", raw_database)
 
-        database.increment_counter("dummy-dir", "clones")
-        self.assertEqual(1, database.get("dummy-dir")["mirror-updates"])
-        self.assertEqual(1, database.get("dummy-dir")["clones"])
-        self.assertEqual(0, database.get("dummy-dir")["updates"])
+        database.increment_counter(repo_abs_path, "mirror-updates")
+        self.assertEqual(1, database.get(repo_abs_path)["mirror-updates"])
+        self.assertEqual(0, database.get(repo_abs_path)["clones"])
+        self.assertEqual(0, database.get(repo_abs_path)["updates"])
 
-        database.increment_counter("dummy-dir", "updates")
-        self.assertEqual(1, database.get("dummy-dir")["mirror-updates"])
-        self.assertEqual(1, database.get("dummy-dir")["clones"])
-        self.assertEqual(1, database.get("dummy-dir")["updates"])
+        database.increment_counter(repo_abs_path, "clones")
+        self.assertEqual(1, database.get(repo_abs_path)["mirror-updates"])
+        self.assertEqual(1, database.get(repo_abs_path)["clones"])
+        self.assertEqual(0, database.get(repo_abs_path)["updates"])
 
-        database.clear_counters("dummy-dir")
-        self.assertEqual(0, database.get("dummy-dir")["mirror-updates"])
-        self.assertEqual(0, database.get("dummy-dir")["clones"])
-        self.assertEqual(0, database.get("dummy-dir")["updates"])
+        database.increment_counter(repo_abs_path, "updates")
+        self.assertEqual(1, database.get(repo_abs_path)["mirror-updates"])
+        self.assertEqual(1, database.get(repo_abs_path)["clones"])
+        self.assertEqual(1, database.get(repo_abs_path)["updates"])
 
-        old_update_time = database.get("dummy-dir")["last-update-time"]
+        database.clear_counters(repo_abs_path)
+        self.assertEqual(0, database.get(repo_abs_path)["mirror-updates"])
+        self.assertEqual(0, database.get(repo_abs_path)["clones"])
+        self.assertEqual(0, database.get(repo_abs_path)["updates"])
+
+        old_update_time = database.get(repo_abs_path)["last-update-time"]
         time.sleep(0.1)
-        database.save_update_time("dummy-dir")
+        database.save_update_time(repo_abs_path)
         time.sleep(0.1)
-        self.assertNotEqual(old_update_time, database.get("dummy-dir")["last-update-time"])
-        self.assertTrue(database.get_time_since_last_update("dummy-dir") < 10.0)
-        self.assertNotEqual(0.0, database.get_time_since_last_update("dummy-dir"))
+        self.assertNotEqual(old_update_time, database.get(repo_abs_path)["last-update-time"])
+        self.assertTrue(database.get_time_since_last_update(repo_abs_path) < 10.0)
+        self.assertNotEqual(0.0, database.get_time_since_last_update(repo_abs_path))
 
-        # self.assertTrue(database.update_time_reached("dummy-dir"))
-        # self.assertFalse(database.cleanup_time_reached("dummy-dir"))
+        # self.assertTrue(database.update_time_reached(repo_abs_path))
+        # self.assertFalse(database.cleanup_time_reached(repo_abs_path))
 
-        database.remove("dummy-dir")
-        self.assertNotIn("dummy-dir", database.get_all())
-        self.assertEqual(None, database.get("dummy-dir"))
-        self.assertEqual(0.0, database.get_time_since_last_update("dummy-dir"))
+        database.remove(repo_abs_path)
+        self.assertNotIn(repo_abs_path, database.get_all())
+        self.assertEqual(None, database.get(repo_abs_path))
+        self.assertEqual(0.0, database.get_time_since_last_update(repo_abs_path))
 
-        self.assertEqual(None, database.get_url_for_path("dummy-dir"))
+        self.assertEqual(None, database.get_url_for_path(repo_abs_path))
 
     # @mockenv(GITCACHE_DIR="/tmp",
     #          GITCACHE_UPDATE_INTERVAL="-1")
@@ -116,8 +124,8 @@ class GitCacheDatabaseTest(TestCase):
     #     importlib.reload(git_cache.database)
 
     #     database = git_cache.database.Database()
-    #     database.add("http://dummy/git", "dummy-dir")
-    #     self.assertFalse(database.update_time_reached("dummy-dir"))
+    #     database.add("http://dummy/git", repo_abs_path)
+    #     self.assertFalse(database.update_time_reached(repo_abs_path))
 
     # @mockenv(GITCACHE_DIR="/tmp",
     #          GITCACHE_UPDATE_INTERVAL="1000")
@@ -128,13 +136,13 @@ class GitCacheDatabaseTest(TestCase):
     #     importlib.reload(git_cache.database)
 
     #     database = git_cache.database.Database()
-    #     database.add("http://dummy/git", "dummy-dir")
-    #     self.assertFalse(database.update_time_reached("dummy-dir"))
+    #     database.add("http://dummy/git", repo_abs_path)
+    #     self.assertFalse(database.update_time_reached(repo_abs_path))
 
-    #     database.database["dummy-dir"]["last-update-time"] -= 2000
+    #     database.database[repo_abs_path]["last-update-time"] -= 2000
     #     # pylint: disable=protected-access
     #     database._save()
-    #     self.assertTrue(database.update_time_reached("dummy-dir"))
+    #     self.assertTrue(database.update_time_reached(repo_abs_path))
     #     self.assertFalse(database.update_time_reached("non-existant"))
 
     # @mockenv(GITCACHE_DIR="/tmp",
@@ -146,13 +154,13 @@ class GitCacheDatabaseTest(TestCase):
     #     importlib.reload(git_cache.database)
 
     #     database = git_cache.database.Database()
-    #     database.add("http://dummy/git", "dummy-dir")
-    #     self.assertFalse(database.cleanup_time_reached("dummy-dir"))
+    #     database.add("http://dummy/git", repo_abs_path)
+    #     self.assertFalse(database.cleanup_time_reached(repo_abs_path))
 
-    #     database.database["dummy-dir"]["last-update-time"] -= 2000
+    #     database.database[repo_abs_path]["last-update-time"] -= 2000
     #     # pylint: disable=protected-access
     #     database._save()
-    #     self.assertTrue(database.cleanup_time_reached("dummy-dir"))
+    #     self.assertTrue(database.cleanup_time_reached(repo_abs_path))
     #     self.assertFalse(database.cleanup_time_reached("non-existant"))